+++ CHANGELOG.md
@@ -32,9 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update Jest major version
 - Changed RosterHeader 'title' prop to all for elements as well as strings
-
 - Changed senderName to optional in ChatBubble
 - Moved children inside of a div in ChatBubble
+- Changed `MeetingManager` to strictly enforce `DevicePermissionStatus` type.
 
 ### Removed